make shortener only shortener
