+++ internal/handlers/handlers.go
@@ -4,6 +4,8 @@ import (
 	"encoding/json"
 	"io"
 	"net/http"
+	"net/url"
+	"os"
 
 	"github.com/gin-gonic/gin"
 	"github.com/khodya/go-urlshortener/internal/shortener"
@@ -18,6 +20,20 @@ const (
 	DefaultBaseURL = "http://localhost:8080"
 )
 
+var baseURL url.URL
+
+func init() {
+	baseURL = parseBaseURL()
+}
+
+func parseBaseURL() url.URL {
+	baseURL, err := url.Parse(os.Getenv(BaseURLEnvName))
+	if err != nil || (*baseURL == url.URL{}) {
+		baseURL, _ = url.Parse(DefaultBaseURL)
+	}
+	return *baseURL
+}
+
 func Fold(c *gin.Context) {
 	defer c.Request.Body.Close()
 	url, err := io.ReadAll(c.Request.Body)
@@ -25,7 +41,8 @@ func Fold(c *gin.Context) {
 		c.Status(400)
 		return
 	}
-	c.String(http.StatusCreated, "%s", shortener.Encode(url))
+	shortURL := composeURL(baseURL, shortener.Encode(url))
+	c.String(http.StatusCreated, "%s", shortURL)
 }
 
 func Unfold(c *gin.Context) {
@@ -50,9 +67,16 @@ func Shorten(c *gin.Context) {
 		c.JSON(http.StatusBadRequest, struct{}{})
 		return
 	}
+	shortURL := shortener.Encode([]byte(requestBody.URLText))
 	c.IndentedJSON(http.StatusCreated, struct {
 		Result string `json:"result"`
 	}{
-		Result: shortener.Encode([]byte(requestBody.URLText)),
+		Result: shortURL,
 	})
+}
+
+func composeURL(baseURL url.URL, path string) string {
+	url := baseURL
+	url.Path = path
+	return url.String()
 }
